requirements.txt
google-adk==0.5.0
httpx==0.28.1
python-dotenv==1.1.0
rtree